3.0.0
This version changes the way state snapshots are passed from one
observer to the next.

Previously, a set of observers would all receive the same state
snapshot, with dispatches (state updates) not propagating until the
entire observer set was exhausted, each subsequent observer effectively
using a stale snapshot.

Thanks to @someonewithpc, this version ensures that each observer,
regardless of it being registered as part of an observer set (e.g., [a,
b, c]) or as a single observer (e.g., [a]), observes the most recent
state snapshot.